backend/api_foodgram/requirements.txt
asgiref==3.7.2
astroid==2.15.6
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
defusedxml==0.7.1
dill==0.3.7
Django==4.2.5
django-filter==23.2
django-stubs==4.2.4
django-stubs-ext==4.2.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.4
mccabe==0.7.0
mypy==1.5.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pybase64==1.3.0
pycparser==2.21
PyJWT==2.8.0
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
types-pytz==2023.3.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
psycopg2-binary==2.9.3
